requirements.txt
# Install with: pip install -r requirements.txt

# =============================================================================
# Core Dependencies
# =============================================================================
numpy>=1.24.0
pyproj>=3.6.0  # Coordinate transforms
pyyaml>=6.0    # Configuration files
easydict>=1.10   # Attribute dict access
optuna>=3.0.0    # Hyperparameter optimization
optuna-integration>=3.0.0 # Optuna integrations

# =============================================================================
# Deep Learning (PyTorch Stack)
# =============================================================================
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0
timm>=0.9.0       # Vision transformers
einops>=0.7.0     # Tensor operations

# =============================================================================
# Data Storage & Processing
# =============================================================================
# Data Storage (LMDB for training datasets)
lmdb>=1.4.0       # Key-value storage with multiprocessing support
pandas>=2.0.0

# =============================================================================
# Scene Generation Dependencies (Migrated from Geo2SigMap)
# =============================================================================
osmnx>=2.0.0      # OpenStreetMap data
opencv-python>=4.8.0 # Fast image processing
shapely>=2.0.0    # Geometric operations
rasterio>=1.3.0   # Geospatial raster data
open3d>=0.17.0    # 3D geometry processing
triangle          # Delaunay triangulation
pyvista>=0.43.0   # 3D visualization and mesh processing
geopandas>=0.14.0 # Geospatial data processing
tqdm>=4.65.0      # Progress bars
laspy[lazrs]      # LiDAR file processing
pdal              # Point data abstraction library
plyfile           # PLY file processing
requests>=2.31.0  # HTTP requests
scipy>=1.11.0     # Scientific computing (for interpolation)
importlib_resources>=6.0.0  # Resource file access
networkx>=3.1
rtree>=1.0.1
scipy>=1.10.0
trimesh>=4.0.0        # Mesh processing

# =============================================================================
# Experiment Tracking & Logging
# =============================================================================
wandb>=0.15.0
comet_ml>=3.31.0
rich>=13.0.0

# =============================================================================
# Testing
# =============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0     # Parallel test execution
pytest-timeout>=2.1.0
pytest-sugar>=0.9.7     # Better test output
pytest-html>=3.2.0      # HTML test reports

# =============================================================================
# Web Interface (Streamlit)
# =============================================================================
streamlit>=1.29.0
plotly>=5.18.0
folium>=0.15.0
streamlit-folium>=0.15.0
pydeck>=0.8.0
Pillow>=10.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0  # PDF export

# =============================================================================
# Optional: Sionna Ray Tracing (Large ~500MB, requires CUDA)
# =============================================================================
sionna>=0.16.0
tensorflow>=2.15.0  # Required by Sionna

# =============================================================================
# Geometric Deep Learning (escnn for rotation equivariance)
# =============================================================================
# Note: lie-learn must be installed from git first:
#   pip install git+https://github.com/AMLab-Amsterdam/lie_learn.git
# Then: LDFLAGS="-L/opt/homebrew/opt/libomp/lib" CPPFLAGS="-I/opt/homebrew/opt/libomp/include" pip install escnn
escnn>=1.0.0
